Start work on moving timers to 64bit ints

+++ src/settings.h
@@ -5,7 +5,7 @@
 typedef struct {
   bool timers_start_auto; // Should timers start immediately when you add them?
   TimerVibration timers_vibration; // Default timer vibration pattern
-  uint16_t timers_duration; // Default timer duration
+  uint32_t timers_duration; // Default timer duration
   bool timers_hours; // Use hours in timers?
   bool show_clock; // Display a clock row in the main screen?
 } Settings;
