+++ src/windows/win-duration.h
@@ -2,7 +2,7 @@
 
 #include "../timer.h"
 
-typedef void (*DurationCallback)(uint16_t duration);
+typedef void (*DurationCallback)(uint32_t duration);
 
 void win_duration_init(void);
 void win_duration_show(uint16_t duration, DurationCallback callback);
